backend/requirements.txt
Flask==2.3.2
Flask-PyMongo==2.3.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
pymongo==4.4.0
python-dotenv==1.0.0
bcrypt==4.0.1
gunicorn==20.1.0
pytest==7.3.1
pytest-flask==1.2.0
flask-socketio==5.3.4
dnspython==2.3.0
email-validator==2.0.0
marshmallow==3.19.0 